EDIT - server auth config

+++ server/app/infrastructure/video_storage_supabase.py
@@ -1,9 +1,8 @@
 from app.core.config import init_supabase
 from app.core.interfaces.video_storage_interface import VideoStorageInterface
 from app.domain.models.video import Video
-from supabase import create_client, Client
 
-supabase: Client = init_supabase()
+supabase = init_supabase()
 
 class SupabaseVideoStorage(VideoStorageInterface):
     def upload_video(self, video: Video) -> Video:
@@ -12,14 +11,14 @@ class SupabaseVideoStorage(VideoStorageInterface):
             raise Exception(response.error.message)
         return Video(**response.data[0])
 
-    def get_videos(self) -> list[Video]:
-        response = supabase.table('videos').select('*').execute()
+    def get_videos(self, user_id: int) -> list[Video]:
+        response = supabase.table('videos').select('*').eq('user_id', user_id).execute()
         if response.error:
             raise Exception(response.error.message)
         return [Video(**video) for video in response.data]
 
-    def delete_video(self, video_id: int) -> Video:
-        response = supabase.table('videos').delete().eq('id', video_id).execute()
+    def delete_video(self, video_id: int, user_id: int) -> Video:
+        response = supabase.table('videos').delete().eq('id', video_id).eq('user_id', user_id).execute()
         if response.error:
             raise Exception(response.error.message)
         if not response.data:
